replace deps to sqlite
