+++ host/clitest/connTest.go
