+++ host/clitest/setuptaskTest.go
