+++ host/persistence/cassandra/cassandra_tool_updateTask_test.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package tests
+package cassandra
 
 import (
 	"log"
