+++ host/persistence/dynamodb/dynamodb_persistence_test.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package tests
+package dynamodb
 
 import (
 	"testing"
